+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+Nothing
+
+## [0.1.2] - 2016-04-22
 ### Changed
 - Throw error if no `Promise` available on invocation, instead of warning on at
 first invocation
@@ -28,7 +31,8 @@ first invocation
 Initial release
 
 
-[Unreleased]: https://github.com/SimenB/wait-until-promise/compare/v0.1.1...HEAD
+[Unreleased]: https://github.com/SimenB/wait-until-promise/compare/v0.1.2...HEAD
+[0.1.2]: https://github.com/SimenB/wait-until-promise/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/SimenB/wait-until-promise/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/SimenB/wait-until-promise/compare/v0.0.2...v0.1.0
 [0.0.2]: https://github.com/SimenB/wait-until-promise/compare/6784c668df6601aa00743ede13fba40526ea09b1...v0.0.2